requirements.txt
pandas
openpyxl
gender-guesser
python-codicefiscale
